[Glint/Sema] Refactor Convert
Also fix some bugs with unary prefix operators :eyes:

This gets three langtest tests passing that were failing before, and I
observe no regressions in any of the other testing frameworks.
